# medical-project-backend/backend/app/static/openapi.yaml
openapi: 3.1.0
info:
  title: Patient Management API
  version: "1.2.0"
  description: >
    API for managing patients, including creating new patients, updating existing patient records (with version history), medical history, medical visits, and retrieving detailed patient information.
servers:
  - url: https://8dj6t53su8wv.share.zrok.io
    description: Local development server
...
paths:
...
  /api/mfa/setup:
    post:
      tags:
        - Authentication
      summary: Set up MFA for the authenticated user
      description: >
        Generates a new TOTP secret and backup codes for the authenticated user.
        Returns the secret and a provisioning URI for QR code generation.
      security:
        - BearerAuth: []
      responses:
        '200':
          description: MFA setup successful
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MFASetupResponse'
        '401':
          description: Unauthorized - Invalid or missing token
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/mfa/enable:
    post:
      tags:
        - Authentication
      summary: Enable MFA for the authenticated user
      description: >
        Enables MFA for the authenticated user after verifying a TOTP code.
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/MFACodeRequest'
      responses:
        '200':
          description: MFA enabled successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: "success"
                  message:
                    type: string
                    example: "MFA enabled successfully"
        '400':
          description: Bad Request - Invalid code or MFA not set up
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Unauthorized - Invalid or missing token
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/mfa/disable:
    post:
      tags:
        - Authentication
      summary: Disable MFA for the authenticated user
      description: >
        Disables MFA for the authenticated user after verifying a TOTP code.
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/MFACodeRequest'
      responses:
        '200':
          description: MFA disabled successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: "success"
                  message:
                    type: string
                    example: "MFA disabled successfully"
        '400':
          description: Bad Request - Invalid code or MFA not enabled
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Unauthorized - Invalid or missing token
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/mfa/verify:
    post:
      tags:
        - Authentication
      summary: Verify an MFA code and obtain a full access token
      description: >
        Verifies a TOTP code for MFA and, if successful, issues a full access token.
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/MFACodeRequest'
      responses:
        '200':
          description: MFA verification successful
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LoginResponse'
              examples:
                SuccessResponse:
                  value:
                    status: "Success"
                    token: "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9..."
                    user:
                      user_id: "99febb18-8a4c-40b6-942f-df3c60d522dd"
                      user_type: "PATIENT"
                      name: "John Doe"
                      medical_insurance_id: "INS123456"
                      requires_mfa: false
        '400':
          description: Bad Request - Invalid code or token not for MFA verification
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Unauthorized - Invalid or missing token
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/mfa/status:
    get:
      tags:
        - Authentication
      summary: Get MFA status for the authenticated user
      description: >
        Returns the current MFA configuration status for the authenticated user.
      security:
        - BearerAuth: []
      responses:
        '200':
          description: MFA status retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MFAStatusResponse'
              examples:
                NotConfigured:
                  summary: MFA not configured
                  value:
                    status: "success"
                    mfa_configured: false
                    mfa_enabled: false
                ConfiguredButDisabled:
                  summary: MFA configured but disabled
                  value:
                    status: "success"
                    mfa_configured: true
                    mfa_enabled: false
                    configured_at: "2023-09-15T14:30:00Z"
                    last_modified: "2023-09-15T14:35:00Z"
                Enabled:
                  summary: MFA enabled
                  value:
                    status: "success"
                    mfa_configured: true
                    mfa_enabled: true
                    configured_at: "2023-09-15T14:30:00Z"
                    last_modified: "2023-09-15T14:35:00Z"
        '401':
          description: Unauthorized - Invalid or missing token
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token obtained from /api/auth/login
  schemas:
...
    ErrorResponse:
      type: object
      description: Schema for error responses.
      properties:
        error:
          type: string
          example: "Invalid foreign key reference."
      required:
        - error
      additionalProperties: false
...
    MFASetupResponse:
      type: object
      description: Response schema for MFA setup.
      properties:
        status:
          type: string
          example: "success"
        secret:
          type: string
          example: "JBSWY3DPEHPK3PXP"
        backup_codes:
          type: array
          items:
            type: string
          example: ["ABCD1234", "EFGH5678"]
        provisioning_uri:
          type: string
          example: "otpauth://totp/MedicalAPI:user@example.com?secret=JBSWY3DPEHPK3PXP&issuer=MedicalAPI"
      required:
        - status
    MFACodeRequest:
      type: object
      description: Request schema for MFA code verification.
      properties:
        code:
          type: string
          example: "123456"
      required:
        - code
    MFAStatusResponse:
      type: object
      description: Response schema for MFA status.
      properties:
        status:
          type: string
          example: "success"
        mfa_configured:
          type: boolean
          example: true
        mfa_enabled:
          type: boolean
          example: true
        configured_at:
          type: string
          format: date-time
          example: "2023-09-15T14:30:00Z"
        last_modified:
          type: string
          format: date-time
          example: "2023-09-15T14:35:00Z"
      required:
        - status
        - mfa_configured
        - mfa_enabled
    LoginResponse:
      type: object
      description: Schema for login responses.
      properties:
        status:
          type: string
          example: "Success"
        token:
          type: string
          example: "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9..."
        temp_token:
          type: string
          example: "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9..."
        user:
          type: object
          properties:
            user_id:
              type: string
              format: uuid
              example: "99febb18-8a4c-40b6-942f-df3c60d522dd"
            user_type:
              type: string
              example: "PATIENT"
            name:
              type: string
              example: "John Doe"
            medical_insurance_id:
              type: string
              example: "INS123456"
            requires_mfa:
              type: boolean
              example: false
      required:
        - status
...